Fixes for refactorisation
