+++ sktime_dl/deeplearning/tests/test_regressors.py
@@ -7,6 +7,7 @@ from sktime.forecasting.compose import RecursiveTimeSeriesRegressionForecaster
 from sktime.forecasting.model_selection import temporal_train_test_split
 
 from sktime_dl.deeplearning import MLPRegressor, MCDCNNRegressor, CNTCRegressor,LSTMFCNRegressor
+
 from sktime_dl.utils.model_lists import (SMALL_NB_EPOCHS,
                                          construct_all_regressors)
