+++ sktime_dl/networks/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    "CNNNetwork",
+]
+
+from sktime_dl.networks._cnn import CNNNetwork
