+++ sktime_dl/regression/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    "CNNRegressor",
+]
+
+from sktime_dl.regression._cnn import CNNRegressor
